requirements.txt
Flask
pyodbc
azure-storage-blob
gunicorn